Test predictions list with test client, fix bug
If no tournament exists, we would not get the 404 as i
hoped, because last() will be None instead of raising
DoesNotExist. So we got an AttributeError on prediction_set.

I put an ipdb breakpoint in the view code to find the bug.
It would be good to introduce ipdb to the more novice readers
at some point.