Test submitting the prediction form, using Webtest
Django's TestCase is unsuitable for testing form submit. Yes, it's
possible to craft a POST request, but it's not guaranteed that a
form will actually submit the same data.

Webtest allows user-like interaction by clicking on a form and
submitting it. In my opinion this is a must for comprehensive
testing.

In this commit we introduce the third parameter to self.assert...()
calls: this is the message that will be displayed on test failure.
Use it to describe why your test fails. This is a better place to
explain your test than a comment line.

As we require a new dependency (django-webtest), we need a place to
put these dependencies in: requirements.txt. Later, we might make
this more sophisticated to allow version pinning and upgrading
(pip-tools is a nice, simple tool for this).